chore: migrate the sitemap addon to Typescript
