+++ packages/addons/sitemap/admin/components/List/Collection/index.tsx
@@ -26,9 +26,9 @@ const ListComponent = (props) => {
     return null;
   }
 
-  items.map((item, key) => {
-    item.get('languages').map((langItem, langKey) => {
-      const formattedItem = {};
+  items.forEach((item, key) => {
+    item.get('languages').forEach((langItem, langKey) => {
+      const formattedItem: any = {};
       formattedItem.name = key;
       formattedItem.langcode = langKey;
       formattedItem.pattern = langItem.get('pattern');
@@ -41,7 +41,7 @@ const ListComponent = (props) => {
   if (items.size === 0) {
     return (
       <EmptyStateLayout
-        content={{ id: 'sitemap.Empty.URLBundles.Description', defaultMessage: 'No URL bundles have been configured yet.' }}
+        content={formatMessage({ id: 'sitemap.Empty.URLBundles.Description', defaultMessage: 'No URL bundles have been configured yet.' })}
         action={<Button onClick={() => openModal()}>{formatMessage({ id: 'sitemap.Empty.URLBundles.Button', defaultMessage: 'Add the first URL bundle' })}</Button>}
       />
     );
