+++ packages/addons/sitemap/admin/components/Loader/index.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 import { Loader as LoaderComponent } from '@strapi/design-system';
 
 const Loader = () => {
-  const style = {
+  const style: React.CSSProperties = {
     display: 'flex',
     justifyContent: 'center',
-    position: 'absolute',
     height: '100%',
     width: '100%',
+    position: 'absolute',
     backgroundColor: 'rgba(255,255,255, 0.6)',
     zIndex: 1,
     alignItems: 'center',
