+++ packages/addons/sitemap/admin/components/PluginIcon/index.tsx
