+++ packages/addons/sitemap/admin/helpers/getSelectedContentType.ts
