+++ packages/addons/sitemap/admin/permissions.ts
