+++ packages/addons/sitemap/admin/tabs/CollectionURLs/index.tsx
@@ -2,12 +2,17 @@ import React, { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Map } from 'immutable';
 
-import { deleteContentType, discardModifiedContentTypes, onChangeContentTypes, submitModal } from '../../state/actions/Sitemap';
+import {
+  deleteContentType,
+  discardModifiedContentTypes,
+  onChangeContentTypes,
+  submitModal,
+} from '../../state/actions/Sitemap';
 import List from '../../components/List/Collection';
 import ModalForm from '../../components/ModalForm';
 
 const CollectionURLs = () => {
-  const state = useSelector((store) => store.get('sitemap', Map()));
+  const state = useSelector((store: any) => store.get('sitemap', Map()));
   const dispatch = useDispatch();
   const [modalOpen, setModalOpen] = useState(false);
   const [uid, setUid] = useState(null);
@@ -52,7 +57,14 @@ const CollectionURLs = () => {
         modifiedState={state.get('modifiedContentTypes')}
         onSubmit={(e) => handleModalSubmit(e)}
         onCancel={(closeModal) => handleModalClose(closeModal)}
-        onChange={(contentType, lang, key, value) => dispatch(onChangeContentTypes(contentType, lang, key, value))}
+        onChange={
+          (
+            contentType,
+            lang,
+            key,
+            value,
+          ) => dispatch(onChangeContentTypes(contentType, lang, key, value))
+        }
         isOpen={modalOpen}
         id={uid}
         lang={langcode}
